Refactor adversarial into new package including discrim nets
