+++ docs/algorithms/gail.rst
@@ -6,12 +6,16 @@ Implements `Generative Adversarial Imitation Learning <https://arxiv.org/abs/160
 
 API
 ===
-.. autoclass:: imitation.algorithms.adversarial.GAIL
+.. autoclass:: imitation.algorithms.adversarial.gail.GAIL
     :members:
     :inherited-members:
     :noindex:
 
-.. autoclass:: imitation.algorithms.adversarial.AdversarialTrainer
+.. autoclass:: imitation.algorithms.adversarial.gail.DiscrimNetGAIL
+    :members:
+    :noindex:
+
+.. autoclass:: imitation.algorithms.adversarial.common.AdversarialTrainer
     :members:
     :inherited-members:
     :noindex:
