+++ docs/modules/imitation.algorithms.adversarial.rst
@@ -1,0 +1,25 @@
+imitation.algorithms.adversarial package
+========================================
+
+Submodules
+----------
+
+imitation.algorithms.adversarial.common module
+----------------------------------------------
+
+.. automodule:: imitation.algorithms.adversarial.common
+
+imitation.algorithms.adversarial.airl module
+--------------------------------------------
+
+.. automodule:: imitation.algorithms.adversarial.airl
+
+imitation.algorithms.adversarial.gail module
+--------------------------------------------
+
+.. automodule:: imitation.algorithms.adversarial.gail
+
+Module contents
+---------------
+
+.. automodule:: imitation.algorithms.adversarial
