+++ docs/modules/imitation.rewards.rst
@@ -4,12 +4,6 @@ imitation.rewards package
 Submodules
 ----------
 
-imitation.rewards.discrim\_nets module
---------------------------------------
-
-
-.. automodule:: imitation.rewards.discrim_nets
-
 imitation.rewards.reward\_nets module
 -------------------------------------
