+++ src/imitation/algorithms/adversarial/__init__.py
@@ -1,0 +1,1 @@
+"""Adversarial imitation learning algorithms, AIRL and GAIL."""
